requirements.txt
Pillow==10.0.1
deepdanbooru==1.0.2
huggingface-hub==0.22.2
numpy==1.26.4
onnx
onnxruntime==1.17.3
opencv-python==4.9.0.80
pandas==2.2.2
sentence-transformers
torch
transformers
